// src/enums/UNIT_FLAGS.ts
export enum UNIT_FLAGS {
    IS_NONE,
    IS_LAND_UNIT = 0,
    IS_AIR_UNIT = 1,
    IS_CATAPULT = 2,
    IS_SEA_UNIT = 3,
    IS_CRITTER = 4,
    IS_BUILDING = 5,
    IS_SUBMARINE = 6,
    IS_DETECTOR = 7,
    IS_WORKER = 8,
    IS_TANKER = 9,
    IS_TRANSPORT = 10,
    IS_OIL_PLATFORM = 11,
    IS_GOLD_STORAGE = 12,
    IS_CORPSE = 13,
    CAN_GROUND_ATTACK = 14,
    IS_UNDEAD = 15,
    IS_SHORE_BUILDING = 16,
    CAN_CAST_SPELLS = 17,
    IS_WOOD_STORAGE = 18,
    CAN_ATTACK = 19,
    IS_TOWER = 20,
    IS_OIL_PATCH = 21,
    IS_MINE = 22,
    IS_HERO = 23,
    IS_OIL_STORAGE = 24,
    IS_SAPPERS_OR_DEMOLITION_SQUAD = 25,
    IS_COWARD = 26,
    IS_ORGANIC = 27,
}